+++ src/todopipeline/pipeline.clj
@@ -10,7 +10,8 @@
 (ns todopipeline.pipeline
   (:require [lambdacd.server :as server]
             [lambdacd.execution :as execution]
-            [lambdacd.core :as core])
+            [lambdacd.core :as core]
+            [lambdacd.util :as utils])
   (:use [lambdacd.control-flow]
         [todopipeline.steps]))
 
@@ -45,13 +46,23 @@
   ))
 
 
+;; # The Configuration.
+;; This is where you define the run-time configuration of LambdaCD. This configuration is passed on to the
+;; individual build-steps in the `:config`-value of the context and will be used by the infrastructure and
+;; build-steps. :home-dir is the directory where LambdaCD will store all it's internal data that should be persisted
+;; over time, such as the last seen revisions of various git-repositories, the build history and so on.
+
+;; For this demonstration, we don't care too much about persisting anything over a longer timeframe so we just use a
+;; temp-directory
+(def config { :home-dir (utils/create-temp-dir)})
+
 ;; # Some infrastructure
 ;; These definitions serve as the entry-points into lambdacd. we need an initialize-function
 ;; that keeps the pipeline running in background and a ring-handler to be able to access
 ;; the LambdaCD REST-API and the UI.
 
 ;; This function does the work of wiring together everything that's necessary for lambdacd to run
-(def pipeline (core/mk-pipeline pipeline-def))
+(def pipeline (core/mk-pipeline pipeline-def config))
 
 ;; Definitions that are used by the lein-ring plugin to run the application